Truncate at the dispositivo, label from it alone, fix the operating point
Truncation replaces masking. Blanking only the outcome phrase leaves its
whole neighbourhood in the text -- "art. 487, I" against "art. 485",
"condeno a parte autora nas custas" against "condeno o reu", "sucumbencia
reciproca" for a split outcome -- and a character n-gram reads that as
easily as it read the phrase. Each document is now cut at its LAST
dispositivo marker and only the text before the cut reaches any model, the
n-gram baseline included. The last marker, because a decision quotes the
ruling under appeal long before it opens its own. The task becomes
predicting the dispositivo from the report and the reasoning, which is the
long-range task the recurrence was supposed to be for.

Labels come from the extracted dispositivo alone, never from the whole
document -- labelling whole documents is what put "parcialmente
procedente" on an order that merely declined to schedule a hearing. A text
with no dispositivo marker is dropped rather than labelled: it is a
procedural act, not a decision with a hidden outcome. On the 17
hand-annotated documents this reproduces the human label 17/17; on the
benchmark it drops 56% of despachos and interlocutorias, and 90% of that
corpus's acordaos, which turn out to be ementa-only.

Operating point, measured rather than assumed: rho = 3776.27 but the
typical gain on a random vector is 149.86, a spectral concentration of
24.8. Scaling by 0.95/rho therefore leaves the typical gain at 0.038 and
the recurrent drive at 0.9% of the input drive -- the topology is not in
the loop. Raising the gain until the bulk is live puts the dominant mode
at ~24, far outside the echo state property. Per-row normalisation by
in-strength (not the max-row-sum scalar) drops the concentration to 3.4;
that is the operating point recorded in the protocol.

The two leading eigenvalues are near-degenerate because they are the
symmetric and antisymmetric combination of two hemisphere-local modes:
v1+v2 carries 94.4% of its energy on the left soma side, v1-v2 94.4% on
the right. The artifact now carries somaSide so this is checkable.

Memory: random_esn drew int64 indices for a matrix whose CSR build
downcasts them, peaking at 352MB for a 79MB result; int32 brings that to
206MB. Traced rather than guessed, since the same overshoot returns at
batch 64.

+++ experiments/malecns_wifi/src/malecns_wifi/characterize.py
@@ -170,8 +170,10 @@ def random_esn(matrix: sp.csr_matrix, *, seed: int) -> tuple[sp.csr_matrix, dict
     rng = np.random.default_rng(seed)
     n = matrix.shape[0]
     nnz = matrix.nnz
-    rows = rng.integers(0, n, size=nnz, dtype=np.int64)
-    cols = rng.integers(0, n, size=nnz, dtype=np.int64)
+    # int32, not the default int64: at 10M edges each index array is 39MB rather
+    # than 82MB, and the CSR build would downcast them anyway.
+    rows = rng.integers(0, n, size=nnz, dtype=np.int32)
+    cols = rng.integers(0, n, size=nnz, dtype=np.int32)
     data = rng.permutation(matrix.data).astype(np.float32, copy=False)
     esn = sp.csr_matrix((data, (rows, cols)), shape=matrix.shape, dtype=np.float32)
     esn.sum_duplicates()
